+++ packages/backend-canvas2d/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @glissade/backend-canvas2d
 
+## 0.66.0-pre.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @glissade/scene@0.66.0-pre.1
+  - @glissade/core@0.66.0-pre.1
+
 ## 0.66.0-pre.0
 
 ### Patch Changes